UI애니메이션 인스팩터 구축중 타이틀 화면에서 화면의 조작이 가능하다는것을 보여주기 위함

+++ LoopHero/AnimationHandler.h
@@ -7,6 +7,6 @@ class AnimationHandler : public GameNode
 public:
 	virtual ~AnimationHandler() {}
 	virtual void Exec(AnimVariable& animVar) = 0;
-	virtual void AddEvent(float time, AnimVariable& animVar) = 0;
+	virtual void AddEvent(int time, AnimVariable& animVar) = 0;
 };
